feature/IVYPORTAL-18677-Third-party-application-with-permissions - Move all method relative to third party application to new Bean
